Make !mtgcard handle different Array stringify side effects
Previously this would "just work" when sending back a single URL instead of 2 (the majority case). Updating something in the stack must have broken that, because it was sending an array with an empty second string, which was messing up something down the stack and crashing the bot with "can't send an empty message" whenever a user would get a single magic card back.